requirements.txt

requests
pytz
python-dotenv
bs4
beautifulsoup4
pymongo
schedule